+++ src/pytask/__init__.py
@@ -32,9 +32,9 @@ from _pytask.mark_utils import has_mark
 from _pytask.mark_utils import remove_marks
 from _pytask.mark_utils import set_marks
 from _pytask.models import CollectionMetadata
-from _pytask.nodes_utils import FilePathNode
-from _pytask.nodes_utils import Node
-from _pytask.nodes_utils import Task
+from _pytask.nodes import FilePathNode
+from _pytask.nodes import Node
+from _pytask.nodes import Task
 from _pytask.outcomes import CollectionOutcome
 from _pytask.outcomes import count_outcomes
 from _pytask.outcomes import Exit